+++ hooks/fetchMedia/index.mjs
@@ -1,12 +1,21 @@
+/**
+ * fetchMedia
+ *
+ * @param {string} mediaId  The unique numerical identifier assigned to each image attachment in the media library.
+ *
+ * @return {Object} todo.
+ */
+
+// External dependencies.
 import { useSelect } from "@wordpress/data"
 import { store as coreStore } from "@wordpress/core-data"
 
-export function useMedia(id) {
+export function fetchMedia(mediaId) {
   return useSelect(
     select => {
       const { getMedia, isResolving, hasFinishedResolution } = select(coreStore)
 
-      const mediaParameters = [id, { context: "view" }]
+      const mediaParameters = [mediaId, { context: "view" }]
 
       return {
         mediaObj: getMedia(...mediaParameters),
@@ -14,6 +23,6 @@ export function useMedia(id) {
         hasResolvedMedia: hasFinishedResolution("getMedia", mediaParameters)
       }
     },
-    [id]
+    [mediaId]
   )
 }
