--- utils/parseMedia/index.mjs
@@ -1,5 +1,0 @@
-export function parseMedia( media, size ) {
-  return (
-    <img src={ media.media_details.sizes[size].source_url } alt={ media.alt_text } />
-  );
-}
